requirements.txt
accelerate==1.0.1
aioesphomeapi==23.1.1
aiohappyeyeballs==2.3.2
aiohttp==3.9.1
aiosignal==1.3.1
ajsonrpc==1.2.0
alembic==1.13.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
argcomplete==3.2.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asteroid-filterbanks==0.4.0
asttokens==2.4.1
async-interrupt==1.1.1
async-lru==2.0.4
attrs==23.2.0
audioread==3.0.1
av==11.0.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bitarray==2.9.2
bitstring==4.1.4
bleach==6.1.0
blender==1.4
blinker==1.8.2
bottle==0.12.25
build==1.2.1
cachetools==5.5.0
certifi==2023.11.17
cffi==1.16.0
chacha20poly1305-reuseable==0.12.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.1
comtypes==1.4.8
contourpy==1.2.1
cryptography==42.0.2
ctranslate2==4.3.1
cycler==0.12.1
cytoolz==0.12.2
debugpy==1.8.0
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.14
diffusers==0.30.3
docopt==0.6.2
ecdsa==0.18.0
einops==0.8.0
esphome==2024.3.1
esphome-dashboard==20240319.0
esptool==4.7.0
eth-account==0.10.0
eth-hash==0.6.0
eth-keyfile==0.7.0
eth-keys==0.5.0
eth-rlp==1.0.0
eth-typing==4.0.0
eth-utils==3.0.0
eth_abi==5.0.0
executing==2.0.1
fastapi==0.112.2
faster-whisper==1.0.0
fastjsonschema==2.19.1
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
google-auth==2.34.0
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.1
gTTS==2.5.3
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.0
humanfriendly==10.0
HyperPyYAML==1.2.2
icmplib==3.0.4
idna==3.6
ifaddr==0.2.0
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
importlib_resources==6.4.4
intel-openmp==2021.4.0
intelhex==2.3.0
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
julius==0.2.7
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.11
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kconfiglib==13.7.1
kiwisolver==1.4.5
kubernetes==30.1.0
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.3.3
lightning-utilities==0.11.5
llvmlite==0.43.0
lru-dict==1.2.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.21.1
mathutils==3.3.0
matplotlib==3.9.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
noiseprotocol==0.3.1
notebook==7.0.7
notebook_shim==0.2.3
numba==0.60.0
numpy==1.26.3
oauthlib==3.2.2
ollama==0.3.3
omegaconf==2.3.0
onnxruntime==1.18.1
opencv-python==4.10.0.84
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
optuna==3.6.1
orjson==3.10.7
overrides==7.6.0
packaging==23.2
paho-mqtt==1.6.1
pandas==2.2.0
pandocfilters==1.5.1
parsimonious==0.9.0
parso==0.8.3
pillow==10.4.0
platformdirs==4.1.0
platformio==6.1.13
playsound==1.3.0
pooch==1.8.2
posthog==3.6.0
primePy==1.3
proglog==0.1.10
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
pure-eval==0.2.2
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pyelftools==0.30
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
PyOpenGL==3.1.7
pyparsing==3.1.2
pyperclip==1.9.0
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.1.0
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-magic==0.4.27
python-wordpress-xmlrpc==2.3
pytorch-lightning==2.3.3
pytorch-metric-learning==2.5.0
pyttsx3==2.98
pytweening==1.2.0
pytz==2023.3.post1
pyunormalize==15.1.0
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
reedsolo==1.7.0
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rlp==4.0.0
rpds-py==0.17.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
scapy==2.5.0
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
semver==3.0.2
Send2Trash==1.8.2
sentencepiece==0.2.0
setuptools==71.0.4
shellingham==1.5.4
silero==0.4.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.38.2
sympy==1.13.1
tabulate==0.9.0
tbb==2021.13.0
tenacity==9.0.0
tensorboardX==2.6.2.2
terminado==0.18.0
threadpoolctl==3.5.0
tinycss2==1.2.1
tokenizers==0.15.2
toolz==0.12.0
torch==2.5.0+cu118
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.5.0+cu118
torchmetrics==1.4.0.post0
torchvision==0.20.0+cu118
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
transformers==4.39.3
trimesh==4.5.2
typer==0.12.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.25.0
voluptuous==0.14.2
watchfiles==0.24.0
wcwidth==0.2.13
web3==6.14.0
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
whisperx @ git+https://github.com/m-bain/whisperx.git@58f00339af7dcc9705ef49d97a1f40764b7cf555
widgetsnbextension==4.0.9
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zeroconf==0.131.0
zipp==3.20.1
